+++ src/Exception/InvalidNodeIdException.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Gianfriaur\OpcuaPhpClient\Exception;
+
+class InvalidNodeIdException extends OpcUaException
+{
+}